migrate line selector to separate class
